dynamic_tco_lambda_app/requirements.txt
boto3
requests
coralogix_logger